models re-organized
