+++ _ide_helper.php
@@ -3,7 +3,7 @@
 
 /**
  * A helper file for Laravel 5, to provide autocomplete information to your IDE
- * Generated for Laravel 5.8.25 on 2019-06-26 19:04:19.
+ * Generated for Laravel 5.8.25 on 2019-06-26 19:23:24.
  *
  * This file should not be included in your code, only analyzed by your IDE!
  *
